requirements.txt
dlt[bigquery]==0.3.0
requests==2.31.0
functions-framework==3.4.0
google-cloud-bigquery==3.17.0
google-cloud-secret-manager==2.16.0
google-cloud-bigquery-storage>=2.24.0 